move option-bucket to a module
