+++ Cargo.toml
@@ -20,10 +20,9 @@ all-features = true
 default = ["map", "std", "entry"]
 std = ["serde?/std"]
 map = ["hashbrown", "std"]
-entry = ["fixed-map-derive/entry", "option-bucket"]
+entry = ["fixed-map-derive/entry"]
 
 [dependencies]
-option-bucket = { path = "option-bucket", optional = true }
 fixed-map-derive = { version = "0.8.0-alpha.2", path = "fixed-map-derive" }
 hashbrown = { version = "0.12.3", optional = true }
 serde = { version = "1.0.145", optional = true, default_features = false }
@@ -42,6 +41,5 @@ harness = false
 
 [workspace]
 members = [
-  "fixed-map-derive",
-  "option-bucket"
+  "fixed-map-derive"
 ]
