+++ src/storage/entry/boolean.rs
@@ -2,7 +2,7 @@
 
 use crate::storage::entry;
 use crate::storage::BooleanStorage;
-use option_bucket::{NoneBucket, OptionBucket, SomeBucket};
+use entry::option_bucket::{NoneBucket, OptionBucket, SomeBucket};
 
 pub struct VacantEntry<'this, V> {
     key: bool,
