+++ src/storage/entry/option_bucket.rs
@@ -1,3 +1,15 @@
+//! # PRIVATE API
+//!
+//! This API is private, for use only in the `derive(Key)` macro.
+//! Usage for other purposes is not supported, and this API will
+//! not abide by Semver stability guarantees.
+//!
+//! If you find that this API would be useful outside its
+//! application in `fixed-map`, let us know.
+//! We can factor it into its own crate.
+//!
+//! # Option Buckets
+//!
 //! Utility for working with [`Option`s][Option]
 //! in cases where we want mutable access to the value within
 //! and the [`Option`] itself (but not at the same time).
@@ -14,7 +26,7 @@
 //!
 //! Safely implement [`Option::get_or_insert`]
 //! ```
-//! use option_bucket::OptionBucket;
+//! use fixed_map::storage::entry::option_bucket::OptionBucket;
 //!
 //! fn get_or_insert<T>(this: &mut Option<T>, value: T) -> &mut T {
 //!     match OptionBucket::new(this) {
@@ -29,7 +41,7 @@
 //!
 //! Safely implement entry API for [`Option`]
 //! ```
-//! use option_bucket::*;
+//! use fixed_map::storage::entry::option_bucket::*;
 //!
 //! struct OccupiedEntry<'a, T> {
 //!     inner: SomeBucket<'a, T>
@@ -72,53 +84,8 @@
 //!     }
 //! }
 //! ```
-//!
-//! # Unsafe Usage
-//!
-//! This crate uses `unwrap_unchecked` to obtain
-//! mutable references to the inner value of `Some`s,
-//! and to skip `drop` when overwriting `None`s.
 
-#![no_std]
-#![deny(missing_docs)]
-// Enable pedantic lints as warnings so we don't break builds when
-// lints are modified or new lints are added to clippy.
-#![warn(
-    // Enable more useful rustc lints
-    absolute_paths_not_starting_with_crate,
-    elided_lifetimes_in_paths,
-    explicit_outlives_requirements,
-    keyword_idents,
-    macro_use_extern_crate,
-    meta_variable_misuse,
-    missing_copy_implementations,
-    missing_docs,
-    non_ascii_idents,
-    noop_method_call,
-    pointer_structural_match,
-    single_use_lifetimes,
-    trivial_casts,
-    trivial_numeric_casts,
-    unreachable_pub,
-    unused_extern_crates,
-    unused_import_braces,
-    unused_lifetimes,
-    unused_macro_rules,
-    unused_qualifications,
-    unused_tuple_struct_fields,
-    variant_size_differences,
-    // Enable pedantic clippy lints
-    clippy::pedantic,
-    // Useful clippy lints for no_std support
-    clippy::std_instead_of_core,
-    clippy::std_instead_of_alloc,
-    clippy::alloc_instead_of_core,
-    // Useful clippy lint for unsafe code
-    clippy::undocumented_unsafe_blocks,
-    // Always require `unsafe` block when using `unsafe`
-    // operations, even in `unsafe` functions
-    unsafe_op_in_unsafe_fn,
-)]
+#![allow(unsafe_code)]
 // `clippy::pedantic` exceptions
 #![allow(clippy::should_implement_trait, clippy::must_use_candidate)]
 
@@ -169,7 +136,7 @@ impl<'a, T> SomeBucket<'a, T> {
     /// # Examples
     ///
     /// ```
-    /// # use option_bucket::SomeBucket;
+    /// # use fixed_map::storage::entry::option_bucket::SomeBucket;
     ///
     /// let mut text: Option<String> = Some("Hello, world!".to_string());
     /// let some = SomeBucket::new(&mut text).unwrap();
@@ -191,7 +158,7 @@ impl<'a, T> SomeBucket<'a, T> {
     /// # Examples
     ///
     /// ```
-    /// # use option_bucket::SomeBucket;
+    /// # use fixed_map::storage::entry::option_bucket::SomeBucket;
     ///
     /// let mut text: Option<String> = Some("Hello, world!".to_string());
     /// let mut some = SomeBucket::new(&mut text).unwrap();
@@ -212,7 +179,7 @@ impl<'a, T> SomeBucket<'a, T> {
     /// # Examples
     ///
     /// ```
-    /// # use option_bucket::SomeBucket;
+    /// # use fixed_map::storage::entry::option_bucket::SomeBucket;
     ///
     /// let mut text: Option<String> = Some("Hello, world!".to_string());
     /// let some = SomeBucket::new(&mut text).unwrap();
@@ -222,7 +189,7 @@ impl<'a, T> SomeBucket<'a, T> {
     /// ```
     ///
     /// ```compile_fail
-    /// # use option_bucket::SomeBucket;
+    /// # use fixed_map::storage::entry::option_bucket::SomeBucket;
     ///
     /// let mut text: Option<String> = Some("Hello, world!".to_string());
     /// let some = SomeBucket::new(&mut text).unwrap();
@@ -243,7 +210,7 @@ impl<'a, T> SomeBucket<'a, T> {
     /// # Examples
     ///
     /// ```
-    /// # use option_bucket::SomeBucket;
+    /// # use fixed_map::storage::entry::option_bucket::SomeBucket;
     ///
     /// let mut x = Some(2);
     /// let mut some = SomeBucket::new(&mut x).unwrap();
@@ -260,7 +227,7 @@ impl<'a, T> SomeBucket<'a, T> {
     /// and consuming this `SomeBucket`.
     ///
     /// ```
-    /// # use option_bucket::SomeBucket;
+    /// # use fixed_map::storage::entry::option_bucket::SomeBucket;
     ///
     /// let mut x = Some(vec![1, 2]);
     /// let some = SomeBucket::new(&mut x).unwrap();
@@ -322,7 +289,7 @@ impl<'a, T> NoneBucket<'a, T> {
     /// operations handling [`drop`]ping the old value
     /// (since we know there was no old value).
     /// ```
-    /// # use option_bucket::NoneBucket;
+    /// # use fixed_map::storage::entry::option_bucket::NoneBucket;
     ///
     /// let mut opt = None;
     /// let mut none = NoneBucket::new(&mut opt).unwrap();
@@ -358,7 +325,7 @@ impl<'a, T> OptionBucket<'a, T> {
     /// # Examples
     ///
     /// ```
-    /// use option_bucket::OptionBucket;
+    /// use fixed_map::storage::entry::option_bucket::OptionBucket;
     ///
     /// let mut none: Option<i32> = None;
     /// let none_bucket = OptionBucket::new(&mut none);
@@ -376,5 +343,135 @@ impl<'a, T> OptionBucket<'a, T> {
             // SAFETY: if conditional ensures that `opt` is `None`
             unsafe { OptionBucket::None(NoneBucket::new_unchecked(opt)) }
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    #![allow(clippy::too_many_lines, unreachable_pub)]
+    use super::*;
+
+    #[test]
+    fn test() {
+        struct OccupiedEntry<'a, T> {
+            inner: SomeBucket<'a, T>,
+        }
+        struct VacantEntry<'a, T> {
+            inner: NoneBucket<'a, T>,
+        }
+
+        enum Entry<'a, T> {
+            Vacant(VacantEntry<'a, T>),
+            Occupied(OccupiedEntry<'a, T>),
+        }
+
+        impl<'a, T> VacantEntry<'a, T> {
+            fn insert(self, value: T) -> &'a mut T {
+                self.inner.insert(value)
+            }
+        }
+
+        impl<'a, T> OccupiedEntry<'a, T> {
+            fn get(&self) -> &T {
+                self.inner.as_ref()
+            }
+
+            fn get_mut(&mut self) -> &mut T {
+                self.inner.as_mut()
+            }
+
+            fn into_mut(self) -> &'a mut T {
+                self.inner.into_mut()
+            }
+
+            fn insert(&mut self, value: T) -> T {
+                self.inner.replace(value)
+            }
+
+            fn remove(self) -> T {
+                self.inner.take()
+            }
+        }
+
+        impl<'a, T> Entry<'a, T> {
+            pub fn or_insert(self, default: T) -> &'a mut T {
+                match self {
+                    Entry::Occupied(entry) => entry.into_mut(),
+                    Entry::Vacant(entry) => entry.insert(default),
+                }
+            }
+
+            pub fn or_insert_with<F: FnOnce() -> T>(self, default: F) -> &'a mut T {
+                match self {
+                    Entry::Occupied(entry) => entry.into_mut(),
+                    Entry::Vacant(entry) => entry.insert(default()),
+                }
+            }
+
+            pub fn and_modify<F: FnOnce(&mut T)>(self, f: F) -> Self {
+                match self {
+                    Entry::Occupied(mut entry) => {
+                        f(entry.get_mut());
+                        Entry::Occupied(entry)
+                    }
+                    Entry::Vacant(entry) => Entry::Vacant(entry),
+                }
+            }
+
+            pub fn or_default(self) -> &'a mut T
+            where
+                T: Default,
+            {
+                match self {
+                    Entry::Occupied(entry) => entry.into_mut(),
+                    Entry::Vacant(entry) => entry.insert(Default::default()),
+                }
+            }
+        }
+
+        trait OptionEntry {
+            type Entry<'this>
+            where
+                Self: 'this;
+            fn entry(&mut self) -> Self::Entry<'_>;
+        }
+
+        impl<T> OptionEntry for Option<T> {
+            type Entry<'this> = Entry<'this, T> where Self: 'this;
+            fn entry(&mut self) -> Self::Entry<'_> {
+                match OptionBucket::new(self) {
+                    OptionBucket::Some(inner) => Entry::Occupied(OccupiedEntry { inner }),
+                    OptionBucket::None(inner) => Entry::Vacant(VacantEntry { inner }),
+                }
+            }
+        }
+
+        let mut even: Option<i32> = None;
+
+        for n in [0, 1, 2, 3, 4, 5, 6, 7, 8, 9] {
+            if n % 2 == 0 {
+                even.entry().and_modify(|x| *x += 1).or_insert(1);
+            }
+        }
+
+        match even.entry() {
+            Entry::Occupied(mut entry) => {
+                assert_eq!(entry.get(), &5);
+                assert_eq!(entry.insert(-3), 5);
+                assert_eq!(entry.remove(), -3);
+            }
+            Entry::Vacant(_) => unreachable!(),
+        }
+        assert!(even.is_none());
+
+        let day_hours = 24;
+
+        let mut x = None;
+        x.entry().or_insert_with(|| day_hours * 60 * 60 * 1000);
+        assert_eq!(x, Some(86_400_000));
+
+        let mut y: Option<u32> = None;
+        y.entry().or_default();
+        assert_eq!(y, Some(0));
     }
 }
